requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.29.0
atlastk==0.13.4
click==8.2.0
fastapi==0.115.0
fastapi-pagination==0.13.1
ghp-import==2.1.0
greenlet==3.2.2
h11==0.16.0
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
Markdown==3.8
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tortoise==0.1.1
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.32.0
watchdog==6.0.0
